requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.2
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
